test: assert reference behaviour through the data layer
Drop tests that re-read raw SQL to confirm writes persisted (now
redundant with Postgres as the sole store), rewrite the clone, import,
and import-task tests to assert through the data-layer resource, and
move the fake-builder legacy_id tests to a dedicated test_fake module.
Raw SQL now remains only in the transaction-rollback test, where the
absence of a row is the assertion.
